allow headings to contain a space after the hashes
Previously the algorithm couldn't handle headings like this:

    #no leading space in heading

It could only handle headings like this:

    # heading with a leading space

Add support for the former.

Note: I looked around on the internet to see if leading spaces were
required in the syntax, but I couldn't find anything definitive - and
github renders it fine regardless. So I guess this is legal.