Add vmt_fence_closing_text
There are other editors that create TOCs together with corresponding
marks. Some of them add additional strings to the closing mark, e.g.
for the opening fence mark `TOC` they use `/TOC` for the closing mark.

With vmt_fence_closing_text it is now possible to adapt to this
situation.